Dirty fix sqlite & faketime
The path to libfaketime is wrong, a better way to find it is
necessary. Also, in last linux-s, the sqlite init doesn't seems to
end.

In the meantime of decent solutions we use a blank db from another
computer that doesn't have the sqlite problem and the libfaketime
path is harcoded.