New mode: fix
The playback event or period migth be so old that the production
machine's db (used to pull inventory and config from) might be unusable.

In that case, a given inventory file reflecting the network state at the
time of target event or period can be imported instead of the current
one in production machine. And all the best channels from all stations
available in the waveforms and in the given inventory file could be
binded.

This new mode still miss the function of creating a waveform file with
all stations.